Move staticDom back into utils module.
We could do backflips to avoid depending on jsdom, but callers would have to pass it (or some other compatible impl) into the corpus framework, and then we'd have to document that. It's better just to admit a dependency on it and rely on rollup and other bundlers to do dead code elimination.