Change default node score from 1 to 0.
This keeps it as the identity in light of our change of the score-combination math from  multiplication to addition. It also means "no confidence", which is the most accurate thing we can say if no computation has been done. Note that this is probably subtly different than "We are 100% confident that this node does NOT have this type."

Add sigmoid() function to utilsForFrontend.