Change ruleset() to take an array of rules as a single param, along with coeffs and biases.
This way, the Ruleset is always valid, and the coeffs and biases are easy to update.

Roll together all non-Side rule() params into a single `options` arg for future expandability. Take 'name' as the first of those options.

Start autogenerated rule numbers at 0, just to align with what the language does with array indices.