Rip out conserveScore() for now.
No one has ever used it, though I think it might actually be *more* useful now that we're going neural-net. However, it's taken me several days to get everything working between it and the per-rule subscore preservation, and I still don't see the end. So, for now, I'm tearing it out.

addScoreFor() adding the weighted score for an entire type to one rule's subscore for that type was definitely a bug, leading to double-counting some subscores.