Switch to the Adam optimizer, which is significantly more turn-key.
All you have to mess with is the learning rate to get a pretty good result: no more manually tuning the decay as well. And LR=1 generally does a pretty great job off the bat. I've changed the defaults to reflect this, upping the iteration count as well so a naive user will get a reasonable production value. I'd still run it to 15000 iterations for the final value--why not?

Here are the raw test results that led me to this decision:

Pop-up overlay ruleset:
* Converges (at default 1.e-3 LR) without any fiddling but not as fast as any of my manually tuned SGD runs. How about if we just run it for 15,000 iterations? It beats SGD run to 300 iterations. 6min vs. 16s (plus lots of human fiddling time).
* Crank the LR to 1 and it hits the floor immediately. The best SGD solution I found doesn't match it until i=1700. The SGD then beats it ever so slightly. However, Adam validation performance is significantly better than SGD.
* LR=.1 does pretty much the same as LR=1, just with less oscillation at the beginning and leveling off to a slightly worse value.

Price-wise price ruleset:
* Levels off at 51 with LR=1, i=1800. The optimal SGD result is obscured in the mists of time but looks like about 55. So, Adam does well.
* Dropping the LR down by a factor of 10, to 0.1, still converges by iteration 1800.
* Adam just seems much less sensitive to choices of LR, which is nice.

Remove unused imports as well.