Shuffle training samples before training.
This is held by some to be a good idea, and it doesn't hurt my login-forms work. It doesn't have a clear relationship to convergence speed on that problem, and the resulting per-tag accuracy is identical. Still, having this in place means I no longer need to consider whether to try it every time I have a problem.