Make BAD labeling reliable. Fix #98.
Remove bound-ruleset caching. The caching was to speed in-browser training, which is now deprecated in favor of the CLI fathom-train.

It may still be illuminating to go chase this down properly, as it may represent a latent bug in Fathom's own caching, but the important thing for now is to make the user-facing tools work.