Ignore node/no-missing-require eslint error
Now that we are actually linting the isVisible.js functional test, eslint reports an error:
```
/home/travis/build/mozilla/fathom/test/functional/isVisible.js

  4:68  error  "../../utilsForFrontend" is not found  node/no-missing-require
```
I believe this is because 'require' is relative to the directory in which the script is found, while eslint is checking relative to the current working directory.

This could be resolved by:
* Ignoring the error (easiest)
* Updating 'eslint-plugin-node' to >= v5.1.0 in which we can apply the 'resolvePaths' setting.
  * This requires also updating 'eslint', as otherwise 'make lint' fails
  * Updating 'eslint' updates default rules, for which there are then dozens of new violations throughout the source code to address
  * I vote that considering updating eslint be a follow-up issue, rather than part of fixing #122

In light of these options, I chose the best, easiest one, since it doesn't blow up everything else.