Exclude 'isVisible' and its called functions from coverage
Add '/* istanbul ignore next */' to stop trying to cover the following methods, since they are executed in the 'isVisible.html' test page, and that scope does not have coverage variables defined.
* ancestors
* isDomElement
* isVisible
* toDomElement

This allows coverage to continue to work (ignoring these functions) and keeps the JS functional tests in the same stage as the rest of the JavaScript tests in TravisCI.

It should be noted that none of these functions previously had test coverage, which is part of why this approach is feasible.