Bundle the mjs files rather than the js ones.
This solves the "default is not exported by side.mjs" errors emitted by rollup, which were due to it treating the js files as CommonJS (which they were), with default rather than named exports. The files do not have default exports, however, which made it grumpy.

The ./* I removed doesn't make any difference, at least now, but less is more.