Remove wu dependency. Close #99.
This removes all the mysterious references to `document` from the `make bundle` output. Those were causing errors when we tried to run it as part of Firefox (see https://github.com/erikrose/gecko-dev/pull/6).

As a bonus, this drops the unminified bundle size (from `make bundle`) from 237K to 88K.

We may make the new iteration utils public at some point, but I'm not ready to commit to that right this moment.