Make utils accessible to consumers of the bundle.
Make utils accessible through the object model for environments in which you can't just reach for the utils file.

We needed to upgrade rollup to understand the `export * as` syntax. As a result, we had to make a few `name` tweaks to the rollup config. Note: we assume backend utils are not useful to consumers of the bundle; they'd be running Node and readily able to import by file.

We can treeshake now without removing util routines.