Fix linter errors.
There are some new ones since we updated the linter and thus the `eslint:recommended` set of rules from which we inherit.

Turn off no-prototype-builtins since we never use builtins on untrusted objects and it's not worth the added syntactical noise.

Crank up the node engine version because the linter says "Async functions are not supported until Node.js 7.6.0." and "Exponential operators are not supported until Node.js 7.0.0.".

Stop ignoring coverage/, which is no longer generated, and the rollup config, which has the wrong extension to be paid attention to.