Don't reinstall Python doc tooling every time we `make doc`.
I had a path to sphinx-js's `__init__.py` in the makefile, but the path was faulty: it did not include the "python3.7" folder that at least my installation creates. Since we do not know what version of Python is on a developer's machine, we cannot know the name of that interstitial folder. So, we fall back to a dotfile strategy as with .npm_installed.