Add `againstElement()`, which lets us efficiently classify only a single element.
Considered making a workalike to `dom()` called `element()` or `one()` (and aliasing `dom()` to `all()`), but then you'd need to change a ruleset every time you wanted to use it differently. That should be the application's decision, not the ruleset writer's. And, of immediate practical value, any ruleset using `element()` in production would need a clumsy switch to use `dom()` in training.