Split the cli and docs targets into their own makefiles. cli and fathom targets all work now.
docs and cli now have their own venvs, which adds efficiency when building the docs in a separate Travis job and heads off any future problems in which a cli requirement conflicts with a docs one. We should merge them again, however, if we ever use sphinx to document the CLI tools.

Bring .gitignore up to date with the new locations of things.