Use an in-project venv for cli every time. Fix #225.
Calling just `pytest` doesn't appear to find the correct pytest, even when PATH is explicitly exported. (I deduce this from the fact that it doesn't see the libs, like sklearn, installed in the venv.) However, executing a shell script that prints out $PATH shows that PATH did, in fact, get passed down to the subshell. If I explicitly set the env var PATH, it runs the correct pytest and finds the libs.