Switch to a multithreaded HTTP server for speed.
This should avoid Vectorizer timeouts as we used to have with standalone fathom-serve, as documented in #67.

As a bonus, we no longer need to keep separate track of the server thread, as `server_close()` takes care of joining all the threads in the pool in ThreadingHTTPServer. Is it a problem that we don't join the master thread? I don't think so, as it's not a daemon thread, so Python itself should wait for it to join before the main process terminates.

We don't actually factor up the server between this and fathom-serve, because the ThreadingHTTPServer symbol is about the only thing in common. We want this one to be silent. In fact, unsilencing it seems to keep requests from being served.