Merge automatic vectorization into fathom-train.
This way, instead of adding fathom-vectorize, yet another tool, we eliminate 2 tools: fathom-list and fathom-serve. We also remove the need to have 3 terminal tabs open, running yarn watch, yarn browser, and fathom-serve. We remove the error-prone hardlinking of the ruleset into FathomFox, which breaks when git changes to a new branch with a different rulesets.js. We eliminate the possibility of forgetting to revectorize after changing a ruleset or samples. And finally, we pave the way to dramatically simplify our teaching and documentation.

I tried to hew to the CLI design of the previous version of the trainer to keep things familiar. We can explain it as "Where you used to pass in a vector file, now feel free to pass in a directory of samples instead." Of course, if you do this, you'll also need to pass in your ruleset and the trainee ID so we can turn the dir into vectors. You can also keep passing in vector files if you want more control in some niche situation, like if you're trying to reproduce results off an old branch. Incidentally, I expect people to use the default locations of the vector caches 99% of the time.

Aggressive caching is in place to remove every possible impediment to using the auto-vectorization. We hash the ruleset and the sample HTML files (though not their resources) and compare them with stored hashes in a cached vector file, all of which takes well under a second. In addition, we cache the built FathomFox artifact so we need run npm or yarn or hit the network only when the user upgrades to a new version of the Fathom CLI tools.