Make tests pass.
I forgot to cast any explicitly specified vector-cache locations to Paths.

But also, unluckily, the test harness's samples make the output formatting code in fathom-train crash because there are never any tag excerpts to show. So we fix that case as well.

Finally, ignore the "2nd line of an `if` condition happens to line up with the contained block" lint error, as making it happy would result either in an overlong line, a linter turd, or not lining up the similar conditions I meant to emphasize the similarity of by lining up in the first place.