Implement the beginnings of a graphing command.
I need some visibility while working on the Pocket features.

We use a consistent scaling such that every char of width is a consistent number of samples, even from feature to feature, so it won't be misleading when we put plusses and minuses in.

Ultimately, I'd like this to not assume that it's the training set being passed in, but that will require some rethinking of our caching architecture.