Add automatic documentation of CLI tools to the main docs.
I tried to strike a balance between them being readable on the commandline and in the Sphinx docs. It would be nice to teach click how to render away ReST syntax.

Since we're now documenting Python modules and thus need to import them, merge the docs and cli venvs. We keep the unified venv in cli/, since that's the one people are used to activating. We could move it to the root level later.

Fix docs Makefile so it no longer needs the venv to be activated. Remove make.bat. There's no way that works anymore.