Add fathom-fox command.
This makes it more convenient for labelers to get FathomFox going in a blank Firefox profile, for those who have the commandline tools installed. Setting up a profile otherwise requires multiple steps and is easy to neglect.

Change fathom_fox_addon() to take a file-like object so we can grab the default ruleset straight out of the zip. Extracting it to a NamedTemporaryFile doesn't work because ZipFile.extract() always extracts the full directory chain of the extracted file.